+++ src/main/java/states/LevelOne.java
@@ -24,8 +24,8 @@ public class LevelOne implements State {
     @Override
     public void init() {
         tileMap = new TileMap(30);
-        tileMap.loadTiles("/tilesets/grasstileset.gif");
-        tileMap.loadMap("/maps/level1-1.map");
+        tileMap.tileLoading.loadTiles("/tilesets/grasstileset.gif");
+        tileMap.mapLoading.loadMap("/maps/level1-1.map");
         tileMap.setPosition(0, 0);
         background.getResource("/backgrounds/levelone.gif");
         berserker = new Berserker(tileMap);
