+++ src/main/java/tilemaps/TileMap.java
@@ -2,31 +2,13 @@ package tilemaps;
 
 import main.GamePanel;
 
-import javax.imageio.ImageIO;
 import java.awt.*;
-import java.awt.image.BufferedImage;
-import java.io.BufferedReader;
-import java.io.InputStream;
-import java.io.InputStreamReader;
 
 public class TileMap {
-    private Tile tile;
-    private double x;
-    private double y;
-    private int xmin;
-    private int ymin;
-    private int xmax;
-    private int ymax;
+    public TileLoader tileLoading;
+    public MapLoader mapLoading;
     private double tween;
-    private int[][] map;
     private int tileSize;
-    private int numRows;
-    private int numCols;
-    private int width;
-    private int height;
-    private BufferedImage bufferedImage;
-    private int numTilesAcross;
-    private Tile[][] tiles;
     private int rowOffset;
     private int colOffset;
     private int numRowsToDraw;
@@ -34,116 +16,43 @@ public class TileMap {
 
     public TileMap(int tileSize) {
         this.tileSize = tileSize;
+        tileLoading = new TileLoader(tileSize);
+        mapLoading = new MapLoader(tileSize);
         numRowsToDraw = GamePanel.HEIGHT / tileSize + 2;
         numColsToDraw = GamePanel.WIDTH / tileSize + 2;
         tween = 0.07;
-    }
-
-    public void loadTiles(String s) {
-        try {
-            bufferedImage = ImageIO.read(getClass().getResourceAsStream(s));
-            numTilesAcross = bufferedImage.getWidth() / tileSize;
-            tiles = new Tile[2][numTilesAcross];
-            BufferedImage subimage;
-            for (int col = 0; col < numTilesAcross; col++) {
-                subimage = bufferedImage.getSubimage(col * tileSize, 0, tileSize, tileSize);
-                tile = new Tile(subimage, 1);
-                tiles[0][col] = new Tile(subimage, tile.NORMAL);
-                subimage = bufferedImage.getSubimage(col * tileSize, tileSize, tileSize, tileSize);
-                tiles[1][col] = new Tile(subimage, tile.BLOCKED);
-            }
-        } catch (Exception e) {
-            e.printStackTrace();
-        }
-    }
-
-    public void loadMap(String s) {
-        try {
-            InputStream in = getClass().getResourceAsStream(s);
-            BufferedReader br = new BufferedReader(new InputStreamReader(in));
-            numCols = Integer.parseInt(br.readLine());
-            numRows = Integer.parseInt(br.readLine());
-            map = new int[numRows][numCols];
-            width = numCols * tileSize;
-            height = numRows * tileSize;
-            xmin = GamePanel.WIDTH - width;
-            xmax = 0;
-            ymin = GamePanel.HEIGHT - height;
-            ymax = 0;
-            String delims = "\\s+";
-            for (int row = 0; row < numRows; row++) {
-                String line = br.readLine();
-                String[] tokens = line.split(delims);
-                for (int col = 0; col < numCols; col++) {
-                    map[row][col] = Integer.parseInt(tokens[col]);
-                }
-            }
-        } catch (Exception e) {
-            e.printStackTrace();
-        }
     }
 
     public int getTileSize() {
         return tileSize;
     }
 
-    public int getx() {
-        return (int) x;
-    }
-
-    public int gety() {
-        return (int) y;
-    }
-
-    public int getWidth() {
-        return width;
-    }
-
-    public int getHeight() {
-        return height;
-    }
-
     public int getType(int row, int col) {
-        int rc = map[row][col];
-        int r = rc / numTilesAcross;
-        int c = rc % numTilesAcross;
-        return tiles[r][c].getType();
+        int rc = mapLoading.map[row][col];
+        int r = rc / tileLoading.numTilesAcross;
+        int c = rc % tileLoading.numTilesAcross;
+        return tileLoading.tiles[r][c].getType();
     }
 
     public void setPosition(double x, double y) {
-        this.x += (x - this.x) * tween;
-        this.y += (y - this.y) * tween;
-        fixBounds();
-        colOffset = (int) -this.x / tileSize;
-        rowOffset = (int) -this.y / tileSize;
+        mapLoading.x += (x - mapLoading.x) * tween;
+        mapLoading.y += (y - mapLoading.y) * tween;
+        mapLoading.fixBounds();
+        colOffset = (int) -mapLoading.x / tileSize;
+        rowOffset = (int) -mapLoading.y / tileSize;
     }
 
     public void draw(Graphics g) {
         for (int row = rowOffset; row < rowOffset + numRowsToDraw; row++) {
-            if (row >= numRows) break;
+            if (row >= mapLoading.numRows) break;
             for (int col = colOffset; col < colOffset + numColsToDraw; col++) {
-                if (col >= numCols) break;
-                if (map[row][col] == 0) continue;
-                int rc = map[row][col];
-                int r = rc / numTilesAcross;
-                int c = rc % numTilesAcross;
-                g.drawImage(tiles[r][c].getImage(), (int) x + col * tileSize, (int) y + row * tileSize, null);
+                if (col >= mapLoading.numCols) break;
+                if (mapLoading.map[row][col] == 0) continue;
+                int rc = mapLoading.map[row][col];
+                int r = rc / tileLoading.numTilesAcross;
+                int c = rc % tileLoading.numTilesAcross;
+                g.drawImage(tileLoading.tiles[r][c].getImage(), (int) mapLoading.x + col * tileSize, (int) mapLoading.y + row * tileSize, null);
             }
-        }
-    }
-
-    private void fixBounds() {
-        if (x < xmin) {
-            x = xmin;
-        }
-        if (y < ymin) {
-            y = ymin;
-        }
-        if (x > xmax) {
-            x = xmax;
-        }
-        if (y > ymax) {
-            y = ymax;
         }
     }
 }
